+++ days/11/day-11.hs
@@ -63,50 +63,46 @@ data ReadPMonkey = ReadPMonkey {
   , rpitems :: [Int]
 } deriving (Show)
 
---- Bing-Chat suggested solution to reading a CSV list of Ints
+--- slightly tweaked Bing-Chat suggested solution to reading a CSV list of Ints
 
 commaSep :: ReadP a -> ReadP [a]
-commaSep p = p `sepBy` (char ',')
+commaSep p = p `sepBy` char ','
+
+commaSpcSep :: ReadP a -> ReadP [a]
+commaSpcSep p = p `sepBy` string ", "
 
 parseInt :: ReadP Int
-parseInt = read <$> many1 (satisfy (`elem` ['0'..'9']))
+parseInt = read <$> many1 (satisfy isDigit)
 
 parseCSV :: ReadP [Int]
 parseCSV = commaSep parseInt
 
----
+parseCommaSpcSV :: ReadP [Int]
+parseCommaSpcSV = commaSpcSep parseInt
 
-parseWhiteSpace :: ReadP Int
-parseWhiteSpace = read <$> many1 (satisfy isSpace)
+---
 
 readPmonkeyID :: ReadP Int
 readPmonkeyID = do
     string "Monkey "
-    -- mID <- parseInt
-    mID <- read <$> many1 (satisfy isDigit)
+    mID <- parseInt
     satisfy (== ':')
+    satisfy (== '\n')
     return mID
 
 readPmonkeyItems :: ReadP [Int]
 readPmonkeyItems = do
-    parseWhiteSpace
+    skipSpaces
     string "Starting items: "
-    items <- parseCSV
+    items <- parseCommaSpcSV
     satisfy (== '\n')
     return items
-
--- metar :: ReadP Report
--- metar = do
---     code <- airport
---     time <- timestamp
---     wind <- windInfo
---     return (Report code time wind)
 
 readPMonkeyData :: ReadP ReadPMonkey
 readPMonkeyData = do
     id <- readPmonkeyID
-    -- items <- readPmonkeyItems
-    return (ReadPMonkey id [42, 42])
+    items <- readPmonkeyItems
+    return (ReadPMonkey id items)
 
 main :: IO ()
 main = do